Changed CommonDTO abstraction name to normal, removed redundant code
