+++ lib/data_service/src/main/kotlin/po/db/data_service/models/CommonDTO.kt
@@ -1,27 +1,34 @@
 package po.db.data_service.models
 
 import org.jetbrains.exposed.dao.LongEntity
-import po.db.data_service.dto.interfaces.DTOEntityMarker
+import po.db.data_service.dto.DTOClass
+import po.db.data_service.dto.interfaces.DTOModelV2
 import po.db.data_service.dto.interfaces.DataModel
-import po.db.data_service.exceptions.ExceptionCodes.NOT_INITIALIZED
-import po.db.data_service.exceptions.InitializationException
+import po.db.data_service.exceptions.ExceptionCodes
+import po.db.data_service.exceptions.OperationsException
 
-
-abstract class CommonDTO<DATA_MODEL, ENTITY>(
-    private val injectedDataModel : DATA_MODEL): DTOEntityMarker<DATA_MODEL, ENTITY>, Cloneable where DATA_MODEL : DataModel, ENTITY : LongEntity{
+abstract class CommonDTO(private val injectedDataModel : DataModel): DTOModelV2, Cloneable {
 
     override var id:Long = 0L
+    private var _dtoModel : DTOClass? = null
+    val dtoModel : DTOClass
+        get(){return  _dtoModel?:
+        throw OperationsException("Trying to access dtoModel property of CommonDTOV2 id :$id while undefined",
+            ExceptionCodes.LAZY_NOT_INITIALIZED) }
+    private var entityDAO : LongEntity? = null
 
-    override val dataModelClassName: String = ""
+    public override fun clone(): DataModel = this.clone()
 
-    public override fun clone(): DATA_MODEL = this.clone()
+    fun toDTO(): DataModel =  this.injectedDataModel
 
-    fun toDTO(): DATA_MODEL =  this.injectedDataModel
-
-    private var _entityDAO : ENTITY? = null
-    val entityDAO : ENTITY
-        get(): ENTITY {
-            return _entityDAO?: throw InitializationException("Trying to access database daoEntity associated with ${this.dataModelClassName}", NOT_INITIALIZED)
+    fun setEntityDAO(entity :LongEntity, dtoModel : DTOClass){
+        this._dtoModel = dtoModel
+        entityDAO = entity
+        if(id != entity.id.value){
+            id = entity.id.value
+            dtoModel.update(toDTO(), entity)
+            //dtoModel.loadChildren<>(entity)
         }
+    }
 
 }
