+++ lib/data_service/src/main/kotlin/po/db/data_service/scope/service/controls/service_registry/DataClasses.kt
@@ -4,7 +4,6 @@ import org.jetbrains.exposed.dao.LongEntity
 import org.jetbrains.exposed.dao.LongEntityClass
 import po.db.data_service.dto.interfaces.DataModel
 import po.db.data_service.models.CommonDTO
-import po.db.data_service.scope.service.controls.ListBuilder
 import kotlin.reflect.KClass
 
 data class ServiceRegistryItem<DATA_MODEL, ENTITY>(
@@ -23,7 +22,7 @@ data class ServiceData<DATA_MODEL, ENTITY>(
 ) where  DATA_MODEL : DataModel, ENTITY : LongEntity
 
 data class DTOData<DATA_MODEL, ENTITY>(
-    val dtoModelClass: KClass<CommonDTO<DATA_MODEL, ENTITY>>,
+    val dtoModelClass: KClass<CommonDTO>,
     val dataModelClass: KClass<DATA_MODEL>,
     val daoEntityModel: LongEntityClass<ENTITY>,
 ) where  DATA_MODEL : DataModel, ENTITY : LongEntity {
@@ -31,7 +30,7 @@ data class DTOData<DATA_MODEL, ENTITY>(
 }
 
 data class ChildDTOData<CHILD_DATA_MODEL, CHILD_ENTITY>(
-    val dtoModelClass: KClass<CommonDTO<CHILD_DATA_MODEL, CHILD_ENTITY>>,
+    val dtoModelClass: KClass<CommonDTO>,
     val dataModelClass: KClass<CHILD_DATA_MODEL>,
     val daoEntityModel: LongEntityClass<CHILD_ENTITY>,
    // val childDTOModels: List<ChildDTOData<*,*>>, // <-Here another list of child in order to create tree like structure
